postprocessing2/requirements.txt
EasySettings
google
lxml
pyenchant
regex
requests
watchdog